fix: resolve lint errors in user tab groups feature
- Move COLOR_MAP to shared/group-colors.ts to fix react-refresh export rule
- Remove unnecessary non-null assertion on draggedTab.userGroupId
- Format useCallback dependency array across multiple lines
- Use nullish coalescing (??) instead of logical OR (||) in separator gating
