+++ src/renderer/src/components/sidebar-constants.ts
@@ -16,4 +16,4 @@ export function clampSidebarWidth(rawWidth: number, viewportWidth: number): numb
   return Math.max(MIN_SIDEBAR_WIDTH, Math.min(rawWidth, max));
 }
 
-export { USER_GROUP_COLORS, type UserGroupColor } from '../../../shared/group-colors';
+export { USER_GROUP_COLORS, COLOR_MAP, type UserGroupColor } from '../../../shared/group-colors';
